add 4 script to test more better. it could test 40 target at once, and when one of them exit will start test another if there is targets to test. and it will kill one which had run more than one hour.
for-klee-test.sh: the main script to test all .bc which find by "-name *.bc"
in the target dir
runklee.sh:  the core script to test by klee, the para could be set in it.
timeout.sh:  the watchdog of test, it kill the test process if it run for a long time
test.sh:    do test.

Signed-off-by: xqx <xqx@server.(none)>